[CDF-21568, CDF-21569] 🪓Move deploy and clean (#562)
* refactor; move methods

* refactor: replace all loaders

* refactor: moved deploy data

* refactor: fixed clean

* refactor: shared base for load and deploy

* refactor: reorg to make clean availabe for deploy

* refactor: fix import

* refactor: moved over deploy and clean calls

* tests: updated tests

* refactor: delete unused round 1

* refactor: delete unused round 2

* refactor; replace custom warnings in deploy

* refactor; replace custom warnings in clean

* refactor; python 3.9 friendly
