+++ cognite_toolkit/_cdf.py
@@ -15,8 +15,7 @@ from dotenv import load_dotenv
 from rich import print
 from rich.panel import Panel
 
-from cognite_toolkit._cdf_tk.commands import BuildCommand, CleanCommand, auth
-from cognite_toolkit._cdf_tk.commands.deploy import DeployCommand
+from cognite_toolkit._cdf_tk.commands import BuildCommand, CleanCommand, DeployCommand, auth
 from cognite_toolkit._cdf_tk.commands.describe import describe_datamodel
 from cognite_toolkit._cdf_tk.commands.dump import dump_datamodel_command
 from cognite_toolkit._cdf_tk.commands.pull import pull_command
