+++ cognite_toolkit/_cdf_tk/commands/__init__.py
@@ -1,4 +1,5 @@
 from .build import BuildCommand
 from .clean import CleanCommand
+from .deploy import DeployCommand
 
-__all__ = ["BuildCommand", "CleanCommand"]
+__all__ = ["BuildCommand", "CleanCommand", "DeployCommand"]
