+++ cognite_toolkit/_cdf_tk/commands/clean.py
@@ -1,7 +1,13 @@
+from __future__ import annotations
+
+import traceback
 from graphlib import TopologicalSorter
 from pathlib import Path
 
 import typer
+from cognite.client.data_classes._base import T_CogniteResourceList
+from cognite.client.exceptions import CogniteAPIError, CogniteNotFoundError
+from cognite.client.utils.useful_types import SequenceNotStr
 from rich import print
 from rich.panel import Panel
 
@@ -15,8 +21,11 @@ from cognite_toolkit._cdf_tk.load import (
     AuthLoader,
     DataSetsLoader,
     DeployResults,
+    ResourceContainerLoader,
     ResourceLoader,
 )
+from cognite_toolkit._cdf_tk.load._base_loaders import T_ID
+from cognite_toolkit._cdf_tk.load.data_classes import ResourceContainerDeployResult, ResourceDeployResult
 from cognite_toolkit._cdf_tk.templates import (
     BUILD_ENVIRONMENT_FILE,
 )
@@ -24,6 +33,8 @@ from cognite_toolkit._cdf_tk.templates.data_classes import (
     BuildEnvironment,
 )
 from cognite_toolkit._cdf_tk.tk_warnings import (
+    LowSeverityWarning,
+    MediumSeverityWarning,
     ToolkitDependenciesIncludedWarning,
     ToolkitNotSupportedWarning,
 )
@@ -33,7 +44,236 @@ from cognite_toolkit._cdf_tk.utils import (
 )
 
 
-class CleanCommand(ToolkitCommand):
+class CleanBaseCommand(ToolkitCommand):
+    def clean_resources(
+        self,
+        loader: ResourceLoader,
+        path: Path,
+        ToolGlobals: CDFToolConfig,
+        dry_run: bool = False,
+        drop: bool = True,
+        drop_data: bool = False,
+        verbose: bool = False,
+    ) -> ResourceDeployResult | None:
+        if not isinstance(loader, ResourceContainerLoader) and not drop:
+            # Skipping silently as this, we will not drop data or delete this resource
+            return ResourceDeployResult(name=loader.display_name)
+        if not loader.support_drop:
+            print(f"  [bold green]INFO:[/] {loader.display_name!r} cleaning is not supported, skipping...")
+            return ResourceDeployResult(name=loader.display_name)
+        elif isinstance(loader, ResourceContainerLoader) and not drop_data:
+            print(
+                f"  [bold]INFO:[/] Skipping cleaning of {loader.display_name!r}. This is a data resource (it contains "
+                f"data and is not only configuration/metadata) and therefore "
+                "requires the --drop-data flag to be set to perform cleaning..."
+            )
+            return ResourceContainerDeployResult(name=loader.display_name, item_name=loader.item_name)
+
+        filepaths = loader.find_files(path)
+
+        # Since we do a clean, we do not want to verify that everything exists wrt data sets, spaces etc.
+        loaded_resources = self._load_files(loader, filepaths, ToolGlobals, skip_validation=True, verbose=verbose)
+        if loaded_resources is None:
+            ToolGlobals.failed = True
+            return None
+
+        # Duplicates should be handled on the build step,
+        # but in case any of them slip through, we do it here as well to
+        # avoid an error.
+        loaded_resources, duplicates = self._remove_duplicates(loaded_resources, loader)
+
+        capabilities = loader.get_required_capability(loaded_resources)
+        if capabilities:
+            ToolGlobals.verify_capabilities(capabilities)
+
+        nr_of_items = len(loaded_resources)
+        if nr_of_items == 0:
+            return ResourceDeployResult(name=loader.display_name)
+
+        existing_resources = loader.retrieve(loader.get_ids(loaded_resources)).as_write()
+        nr_of_existing = len(existing_resources)
+
+        if drop:
+            prefix = "Would clean" if dry_run else "Cleaning"
+            with_data = "with data " if isinstance(loader, ResourceContainerLoader) else ""
+        else:
+            prefix = "Would drop data from" if dry_run else "Dropping data from"
+            with_data = ""
+        print(f"[bold]{prefix} {nr_of_existing} {loader.display_name} {with_data}from CDF...[/]")
+        for duplicate in duplicates:
+            self.warn(LowSeverityWarning(f"Duplicate {loader.display_name} {duplicate}."))
+
+        # Deleting resources.
+        if isinstance(loader, ResourceContainerLoader) and drop_data:
+            nr_of_dropped_datapoints = self._drop_data(existing_resources, loader, dry_run, verbose)
+            if drop:
+                nr_of_deleted = self._delete_resources(existing_resources, loader, dry_run, verbose)
+            else:
+                nr_of_deleted = 0
+            if verbose:
+                print("")
+            return ResourceContainerDeployResult(
+                name=loader.display_name,
+                deleted=nr_of_deleted,
+                total=nr_of_items,
+                dropped_datapoints=nr_of_dropped_datapoints,
+                item_name=loader.item_name,
+            )
+        elif not isinstance(self, ResourceContainerLoader) and drop:
+            nr_of_deleted = self._delete_resources(existing_resources, loader, dry_run, verbose)
+            if verbose:
+                print("")
+            return ResourceDeployResult(name=loader.display_name, deleted=nr_of_deleted, total=nr_of_items)
+        else:
+            return ResourceDeployResult(name=loader.display_name)
+
+    def _delete_resources(
+        self, loaded_resources: T_CogniteResourceList, loader: ResourceLoader, dry_run: bool, verbose: bool
+    ) -> int:
+        nr_of_deleted = 0
+        resource_ids = loader.get_ids(loaded_resources)
+        if dry_run:
+            nr_of_deleted += len(resource_ids)
+            if verbose:
+                print(f"  Would have deleted {self._print_ids_or_length(resource_ids)}.")
+            return nr_of_deleted
+
+        try:
+            nr_of_deleted += loader.delete(resource_ids)
+        except CogniteAPIError as e:
+            self.warn(MediumSeverityWarning(f"Failed to delete {self._print_ids_or_length(resource_ids)}. Error {e}."))
+            if verbose:
+                print(Panel(traceback.format_exc()))
+        except CogniteNotFoundError:
+            if verbose:
+                print(f"  [bold]INFO:[/] {self._print_ids_or_length(resource_ids)} do(es) not exist.")
+        except Exception as e:
+            self.warn(MediumSeverityWarning(f"Failed to delete {self._print_ids_or_length(resource_ids)}. Error {e}."))
+            if verbose:
+                print(Panel(traceback.format_exc()))
+        else:  # Delete succeeded
+            if verbose:
+                print(f"  Deleted {self._print_ids_or_length(resource_ids)}.")
+        return nr_of_deleted
+
+    def _drop_data(
+        self, loaded_resources: T_CogniteResourceList, loader: ResourceContainerLoader, dry_run: bool, verbose: bool
+    ) -> int:
+        nr_of_dropped = 0
+        resource_ids = loader.get_ids(loaded_resources)
+        if dry_run:
+            resource_drop_count = loader.count(resource_ids)
+            nr_of_dropped += resource_drop_count
+            if verbose:
+                self._verbose_print_drop(resource_drop_count, resource_ids, loader, dry_run)
+            return nr_of_dropped
+
+        try:
+            resource_drop_count = loader.drop_data(resource_ids)
+            nr_of_dropped += resource_drop_count
+        except CogniteAPIError as e:
+            if e.code == 404 and verbose:
+                print(f"  [bold]INFO:[/] {len(resource_ids)} {loader.display_name} do(es) not exist.")
+        except CogniteNotFoundError:
+            return nr_of_dropped
+        except Exception as e:
+            self.warn(
+                MediumSeverityWarning(
+                    f"Failed to drop {loader.item_name} from {len(resource_ids)} {loader.display_name}. Error {e}."
+                )
+            )
+            if verbose:
+                print(Panel(traceback.format_exc()))
+        else:  # Delete succeeded
+            if verbose:
+                self._verbose_print_drop(resource_drop_count, resource_ids, loader, dry_run)
+        return nr_of_dropped
+
+    def _verbose_print_drop(
+        self, drop_count: int, resource_ids: SequenceNotStr[T_ID], loader: ResourceContainerLoader, dry_run: bool
+    ) -> None:
+        prefix = "Would have dropped" if dry_run else "Dropped"
+        if drop_count > 0:
+            print(
+                f"  {prefix} {drop_count:,} {loader.item_name} from {loader.display_name}: "
+                f"{self._print_ids_or_length(resource_ids)}."
+            )
+        elif drop_count == 0:
+            verb = "is" if len(resource_ids) == 1 else "are"
+            print(
+                f"  The {loader.display_name}: {self._print_ids_or_length(resource_ids)} {verb} empty, "
+                f"thus no {loader.item_name} will be {'touched' if dry_run else 'dropped'}."
+            )
+        else:
+            # Count is not supported
+            print(
+                f" {prefix} all {loader.item_name} from {loader.display_name}: "
+                f"{self._print_ids_or_length(resource_ids)}."
+            )
+
+    def _load_files(
+        self,
+        loader: ResourceLoader,
+        filepaths: list[Path],
+        ToolGlobals: CDFToolConfig,
+        skip_validation: bool,
+        verbose: bool = False,
+    ) -> T_CogniteResourceList | None:
+        loaded_resources = loader.create_empty_of(loader.list_write_cls([]))
+        for filepath in filepaths:
+            try:
+                resource = loader.load_resource(filepath, ToolGlobals, skip_validation)
+            except KeyError as e:
+                # KeyError means that we are missing a required field in the yaml file.
+                print(
+                    f"[bold red]ERROR:[/] Failed to load {filepath.name} with {loader.display_name}. Missing required field: {e}."
+                    f"[bold red]ERROR:[/] Please compare with the API specification at {loader.doc_url()}."
+                )
+                return None
+            except Exception as e:
+                print(f"[bold red]ERROR:[/] Failed to load {filepath.name} with {loader.display_name}. Error: {e!r}.")
+                if verbose:
+                    print(Panel(traceback.format_exc()))
+                return None
+            if resource is None:
+                # This is intentional. It is, for example, used by the AuthLoader to skip groups with resource scopes.
+                continue
+            if isinstance(resource, loader.list_write_cls) and not resource:
+                self.warn(LowSeverityWarning(f"Skipping {filepath.name}. No data to load."))
+                continue
+
+            if isinstance(resource, loader.list_write_cls):
+                loaded_resources.extend(resource)
+            else:
+                loaded_resources.append(resource)
+        return loaded_resources
+
+    @staticmethod
+    def _print_ids_or_length(resource_ids: SequenceNotStr[T_ID], limit: int = 10) -> str:
+        if len(resource_ids) == 1:
+            return f"{resource_ids[0]!r}"
+        elif len(resource_ids) <= limit:
+            return f"{resource_ids}"
+        else:
+            return f"{len(resource_ids)} items"
+
+    def _remove_duplicates(
+        self, loaded_resources: T_CogniteResourceList, loader: ResourceLoader
+    ) -> tuple[T_CogniteResourceList, list[T_ID]]:
+        seen: set[T_ID] = set()
+        output = loader.create_empty_of(loaded_resources)
+        duplicates: list[T_ID] = []
+        for item in loaded_resources:
+            identifier = loader.get_id(item)
+            if identifier not in seen:
+                output.append(item)
+                seen.add(identifier)
+            else:
+                duplicates.append(identifier)
+        return output, duplicates
+
+
+class CleanCommand(CleanBaseCommand):
     def execute(
         self, ctx: typer.Context, build_dir: str, build_env_name: str, dry_run: bool, include: list[str]
     ) -> None:
@@ -73,7 +313,8 @@ class CleanCommand(ToolkitCommand):
             if type(loader) is DataSetsLoader:
                 self.warn(ToolkitNotSupportedWarning(feature="Dataset clean."))
                 continue
-            result = loader.clean_resources(
+            result = self.clean_resources(
+                loader,
                 build_path / loader_cls.folder_name,
                 ToolGlobals,
                 drop=True,
@@ -91,7 +332,8 @@ class CleanCommand(ToolkitCommand):
                 raise ToolkitCleanResourceError(f"Failure to clean {loader_cls.folder_name} as expected.")
 
         if "auth" in include and (directory := (Path(build_dir) / "auth")).is_dir():
-            result = AuthLoader.create_loader(ToolGlobals, target_scopes="all").clean_resources(
+            result = self.clean_resources(
+                AuthLoader.create_loader(ToolGlobals, target_scopes="all"),
                 directory,
                 ToolGlobals,
                 drop=True,
