+++ cognite_toolkit/_cdf_tk/commands/deploy.py
@@ -1,11 +1,17 @@
+from __future__ import annotations
+
+import re
+import traceback
 from graphlib import TopologicalSorter
 from pathlib import Path
 
 import typer
+from cognite.client.data_classes._base import T_CogniteResourceList
+from cognite.client.exceptions import CogniteAPIError, CogniteDuplicatedError
 from rich import print
 from rich.panel import Panel
 
-from cognite_toolkit._cdf_tk.commands._base import ToolkitCommand
+from cognite_toolkit._cdf_tk.commands.clean import CleanBaseCommand
 from cognite_toolkit._cdf_tk.constants import _RUNNING_IN_BROWSER
 from cognite_toolkit._cdf_tk.exceptions import (
     ToolkitCleanResourceError,
@@ -15,8 +21,18 @@ from cognite_toolkit._cdf_tk.exceptions import (
 from cognite_toolkit._cdf_tk.load import (
     LOADER_BY_FOLDER_NAME,
     AuthLoader,
+    DataLoader,
     DeployResults,
+    Loader,
+    ResourceContainerLoader,
     ResourceLoader,
+)
+from cognite_toolkit._cdf_tk.load.data_classes import (
+    DatapointDeployResult,
+    DeployResult,
+    ResourceContainerDeployResult,
+    ResourceDeployResult,
+    UploadDeployResult,
 )
 from cognite_toolkit._cdf_tk.templates import (
     BUILD_ENVIRONMENT_FILE,
@@ -24,14 +40,18 @@ from cognite_toolkit._cdf_tk.templates import (
 from cognite_toolkit._cdf_tk.templates.data_classes import (
     BuildEnvironment,
 )
-from cognite_toolkit._cdf_tk.tk_warnings.other import ToolkitDependenciesIncludedWarning
+from cognite_toolkit._cdf_tk.tk_warnings.other import (
+    LowSeverityWarning,
+    MediumSeverityWarning,
+    ToolkitDependenciesIncludedWarning,
+)
 from cognite_toolkit._cdf_tk.utils import (
     CDFToolConfig,
     read_yaml_file,
 )
 
 
-class DeployCommand(ToolkitCommand):
+class DeployCommand(CleanBaseCommand):
     def execute(
         self,
         ctx: typer.Context,
@@ -71,6 +91,7 @@ class DeployCommand(ToolkitCommand):
         if len(ordered_loaders) > len(selected_loaders):
             dependencies = [item.folder_name for item in ordered_loaders if item not in selected_loaders]
             self.warn(ToolkitDependenciesIncludedWarning(dependencies))
+        result: DeployResult | None
         if drop or drop_data:
             # Drop has to be done in the reverse order of deploy.
             if drop and drop_data:
@@ -84,7 +105,8 @@ class DeployCommand(ToolkitCommand):
                 if not issubclass(loader_cls, ResourceLoader):
                     continue
                 loader = loader_cls.create_loader(ToolGlobals)
-                result = loader.clean_resources(
+                result = self.clean_resources(
+                    loader,
                     build_path / loader_cls.folder_name,
                     ToolGlobals,
                     drop=drop,
@@ -98,7 +120,8 @@ class DeployCommand(ToolkitCommand):
                     raise ToolkitCleanResourceError(f"Failure to clean {loader_cls.folder_name} as expected.")
 
             if "auth" in include and (directory := (Path(build_dir) / "auth")).is_dir():
-                result = AuthLoader.create_loader(ToolGlobals, target_scopes="all").clean_resources(
+                result = self.clean_resources(
+                    AuthLoader.create_loader(ToolGlobals, target_scopes="all"),
                     directory,
                     ToolGlobals,
                     drop=drop,
@@ -125,9 +148,8 @@ class DeployCommand(ToolkitCommand):
         if "auth" in include and (directory := (Path(build_dir) / "auth")).is_dir():
             # First, we need to get all the generic access, so we can create the rest of the resources.
             result = (
-                AuthLoader
-                .create_loader(ToolGlobals, target_scopes="all_scoped_only")
-                .deploy_resources(directory, **arguments)
+                self.deploy_resources(AuthLoader
+                .create_loader(ToolGlobals, target_scopes="all_scoped_only"), directory, **arguments)
             )  # fmt: skip
             if ToolGlobals.failed:
                 raise ToolkitDeployResourceError("Failure to deploy auth (groups) with ALL scope as expected.")
@@ -137,8 +159,8 @@ class DeployCommand(ToolkitCommand):
                 print("")  # Extra newline
 
         for loader_cls in ordered_loaders:
-            result = loader_cls.create_loader(ToolGlobals).deploy_resources(  # type: ignore[assignment]
-                build_path / loader_cls.folder_name, **arguments
+            result = self.deploy_resources(
+                loader_cls.create_loader(ToolGlobals), build_path / loader_cls.folder_name, **arguments
             )
             if ToolGlobals.failed:
                 if results and results.has_counts:
@@ -155,7 +177,7 @@ class DeployCommand(ToolkitCommand):
             # Last, we create the Groups again, but this time we do not filter out any capabilities
             # and we do not skip validation as the resources should now have been created.
             loader = AuthLoader.create_loader(ToolGlobals, target_scopes="resource_scoped_only")
-            result = loader.deploy_resources(directory, **arguments)
+            result = self.deploy_resources(loader, directory, **arguments)
             if ToolGlobals.failed:
                 raise ToolkitDeployResourceError("Failure to deploy auth (groups) scoped to resources as expected.")
             if result:
@@ -166,3 +188,260 @@ class DeployCommand(ToolkitCommand):
             print(results.uploads_table())
         if ToolGlobals.failed:
             raise ToolkitDeployResourceError("Failure to deploy auth (groups) scoped to resources as expected.")
+
+    def deploy_resources(
+        self,
+        loader: Loader,
+        path: Path,
+        ToolGlobals: CDFToolConfig,
+        dry_run: bool = False,
+        has_done_drop: bool = False,
+        has_dropped_data: bool = False,
+        verbose: bool = False,
+    ) -> DeployResult | None:
+        if isinstance(loader, ResourceLoader):
+            return self._deploy_resources(loader, path, ToolGlobals, dry_run, has_done_drop, has_dropped_data, verbose)
+        elif isinstance(loader, DataLoader):
+            return self._deploy_data(loader, path, ToolGlobals, dry_run, verbose)
+        else:
+            raise ValueError(f"Unsupported loader type {type(loader)}.")
+
+    def _deploy_resources(
+        self,
+        loader: ResourceLoader,
+        path: Path,
+        ToolGlobals: CDFToolConfig,
+        dry_run: bool = False,
+        has_done_drop: bool = False,
+        has_dropped_data: bool = False,
+        verbose: bool = False,
+    ) -> ResourceDeployResult | None:
+        loader.build_path = path
+        filepaths = loader.find_files(path)
+
+        def sort_key(p: Path) -> int:
+            if result := re.findall(r"^(\d+)", p.stem):
+                return int(result[0])
+            else:
+                return len(filepaths)
+
+        # In the build step, the resource files are prefixed a number that controls the order in which
+        # the resources are deployed. The custom 'sort_key' here is to get a sort on integer instead of a default string
+        # sort.
+        filepaths = sorted(filepaths, key=sort_key)
+
+        loaded_resources = self._load_files(loader, filepaths, ToolGlobals, skip_validation=dry_run, verbose=verbose)
+        if loaded_resources is None:
+            ToolGlobals.failed = True
+            return None
+
+        # Duplicates should be handled on the build step,
+        # but in case any of them slip through, we do it here as well to
+        # avoid an error.
+        loaded_resources, duplicates = self._remove_duplicates(loaded_resources, loader)
+
+        if not loaded_resources:
+            return ResourceDeployResult(name=loader.display_name)
+
+        capabilities = loader.get_required_capability(loaded_resources)
+        if capabilities:
+            ToolGlobals.verify_capabilities(capabilities)
+
+        nr_of_items = len(loaded_resources)
+        if nr_of_items == 0:
+            return ResourceDeployResult(name=loader.display_name)
+
+        prefix = "Would deploy" if dry_run else "Deploying"
+        print(f"[bold]{prefix} {nr_of_items} {loader.display_name} to CDF...[/]")
+        # Moved here to avoid printing before the above message.
+        for duplicate in duplicates:
+            self.warn(LowSeverityWarning(f"Skipping duplicate {loader.display_name} {duplicate}."))
+
+        nr_of_created = nr_of_changed = nr_of_unchanged = 0
+        to_create, to_update, unchanged = self.to_create_changed_unchanged_triple(loaded_resources, loader)
+
+        if dry_run:
+            if (
+                loader.support_drop
+                and has_done_drop
+                and (not isinstance(loader, ResourceContainerLoader) or has_dropped_data)
+            ):
+                # Means the resources will be deleted and not left unchanged or changed
+                for item in unchanged:
+                    # We cannot use extents as LoadableNodes cannot be extended.
+                    to_create.append(item)
+                for item in to_update:
+                    to_create.append(item)
+                unchanged.clear()
+                to_update.clear()
+
+            nr_of_unchanged += len(unchanged)
+            nr_of_created += len(to_create)
+            nr_of_changed += len(to_update)
+        else:
+            nr_of_unchanged += len(unchanged)
+
+            if to_create:
+                created = self._create_resources(to_create, loader, verbose)
+                if created is None:
+                    ToolGlobals.failed = True
+                    return None
+                nr_of_created += created
+
+            if to_update:
+                updated = self._update_resources(to_update, loader, verbose)
+                if updated is None:
+                    ToolGlobals.failed = True
+                    return None
+
+                nr_of_changed += updated
+
+        if verbose:
+            self._verbose_print(to_create, to_update, unchanged, loader, dry_run)
+
+        if isinstance(loader, ResourceContainerLoader):
+            return ResourceContainerDeployResult(
+                name=loader.display_name,
+                created=nr_of_created,
+                changed=nr_of_changed,
+                unchanged=nr_of_unchanged,
+                total=nr_of_items,
+                item_name=loader.item_name,
+            )
+        else:
+            return ResourceDeployResult(
+                name=loader.display_name,
+                created=nr_of_created,
+                changed=nr_of_changed,
+                unchanged=nr_of_unchanged,
+                total=nr_of_items,
+            )
+
+    def to_create_changed_unchanged_triple(
+        self,
+        resources: T_CogniteResourceList,
+        loader: ResourceLoader,
+    ) -> tuple[T_CogniteResourceList, T_CogniteResourceList, T_CogniteResourceList]:
+        """Returns a triple of lists of resources that should be created, updated, and are unchanged."""
+        resource_ids = loader.get_ids(resources)
+        to_create, to_update, unchanged = (
+            loader.create_empty_of(resources),
+            loader.create_empty_of(resources),
+            loader.create_empty_of(resources),
+        )
+        try:
+            cdf_resources = loader.retrieve(resource_ids)
+        except Exception as e:
+            self.warn(
+                MediumSeverityWarning(
+                    f"Failed to retrieve {len(resource_ids)} of {loader.display_name}. Proceeding assuming not data in CDF. Error {e}."
+                )
+            )
+            print(Panel(traceback.format_exc()))
+            cdf_resource_by_id = {}
+        else:
+            cdf_resource_by_id = {loader.get_id(resource): resource for resource in cdf_resources}
+
+        for item in resources:
+            cdf_resource = cdf_resource_by_id.get(loader.get_id(item))
+            # The custom compare is needed when the regular == does not work. For example, TransformationWrite
+            # have OIDC credentials that will not be returned by the retrieve method, and thus need special handling.
+            if cdf_resource and (item == cdf_resource.as_write() or loader._is_equal_custom(item, cdf_resource)):
+                unchanged.append(item)
+            elif cdf_resource:
+                to_update.append(item)
+            else:
+                to_create.append(item)
+        return to_create, to_update, unchanged
+
+    def _verbose_print(
+        self,
+        to_create: T_CogniteResourceList,
+        to_update: T_CogniteResourceList,
+        unchanged: T_CogniteResourceList,
+        loader: ResourceLoader,
+        dry_run: bool,
+    ) -> None:
+        print_outs = []
+        prefix = "Would have " if dry_run else ""
+        if to_create:
+            print_outs.append(f"{prefix}Created {self._print_ids_or_length(loader.get_ids(to_create))}")
+        if to_update:
+            print_outs.append(f"{prefix}Updated {self._print_ids_or_length(loader.get_ids(to_update))}")
+        if unchanged:
+            print_outs.append(
+                f"{'Untouched' if dry_run else 'Unchanged'} {self._print_ids_or_length(loader.get_ids(unchanged))}"
+            )
+        prefix_message = f" {loader.display_name}: "
+        if len(print_outs) == 1:
+            print(f"{prefix_message}{print_outs[0]}")
+        elif len(print_outs) == 2:
+            print(f"{prefix_message}{print_outs[0]} and {print_outs[1]}")
+        else:
+            print(f"{prefix_message}{', '.join(print_outs[:-1])} and {print_outs[-1]}")
+
+    def _create_resources(self, resources: T_CogniteResourceList, loader: ResourceLoader, verbose: bool) -> int | None:
+        try:
+            created = loader.create(resources)
+        except CogniteAPIError as e:
+            if e.code == 409:
+                self.warn(LowSeverityWarning("Resource(s) already exist(s), skipping creation."))
+            else:
+                print(f"[bold red]ERROR:[/] Failed to create resource(s).\n{e}")
+                return None
+        except CogniteDuplicatedError as e:
+            self.warn(
+                LowSeverityWarning(
+                    f"{len(e.duplicated)} out of {len(resources)} resource(s) already exist(s). {len(e.successful or [])} resource(s) created."
+                )
+            )
+        except Exception as e:
+            print(f"[bold red]ERROR:[/] Failed to create resource(s).\n{e}")
+            if verbose:
+                print(Panel(traceback.format_exc()))
+            return None
+        else:
+            return len(created) if created is not None else 0
+        return 0
+
+    def _update_resources(self, resources: T_CogniteResourceList, loader: ResourceLoader, verbose: bool) -> int | None:
+        try:
+            updated = loader.update(resources)
+        except Exception as e:
+            print(f"  [bold yellow]Error:[/] Failed to update {loader.display_name}. Error {e}.")
+            if verbose:
+                print(Panel(traceback.format_exc()))
+            return None
+        else:
+            return len(updated)
+
+    def _deploy_data(
+        self,
+        loader: DataLoader,
+        path: Path,
+        ToolGlobals: CDFToolConfig,
+        dry_run: bool = False,
+        verbose: bool = False,
+    ) -> UploadDeployResult | None:
+        filepaths = loader.find_files(path)
+
+        prefix = "Would upload" if dry_run else "Uploading"
+        print(f"[bold]{prefix} {len(filepaths)} data {loader.display_name} files to CDF...[/]")
+        datapoints = 0
+        for filepath in filepaths:
+            try:
+                message, file_datapoints = loader.upload(filepath, ToolGlobals, dry_run)
+            except Exception as e:
+                print(f"  [bold red]Error:[/] Failed to upload {filepath.name}. Error: {e!r}.")
+                print(Panel(traceback.format_exc()))
+                ToolGlobals.failed = True
+                return None
+            if verbose:
+                print(message)
+            datapoints += file_datapoints
+        if datapoints != 0:
+            return DatapointDeployResult(
+                loader.display_name, points=datapoints, uploaded=len(filepaths), item_name=loader.item_name
+            )
+        else:
+            return UploadDeployResult(loader.display_name, uploaded=len(filepaths), item_name=loader.item_name)
